+++ src/main/java/com/munhwa/prj/cart/web/CartController.java
@@ -1,34 +1,35 @@
 package com.munhwa.prj.cart.web;
 
-import java.util.ArrayList;
-import java.util.List;
+import java.util.HashMap;
 import java.util.Map;
 
 import javax.servlet.http.HttpServletRequest;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import com.munhwa.prj.cart.service.CartService;
-import com.munhwa.prj.cart.vo.CartVO;
 import com.munhwa.prj.music.vo.MusicVO;
 
 @Controller
 public class CartController {
 	
-	@Autowired
-	private CartService cartDao;
+//	@Autowired
+//	private CartService cartDao;
 	
-//	@GetMapping("/shop/cart")
-//	public String selectCart(CartVO vo, Model model) {
-//		model.addAttribute("cart", cartDao.selectCart(vo));
-//		return "shop/shop_cart";
-//	}
-	
+
+	//카트 세션 테스트
+	@GetMapping("/cart/test")
+	@ResponseBody
+	public String cart(HttpServletRequest req) {
+		Map<Integer, MusicVO> map = new HashMap<Integer, MusicVO>();
+		req.getSession().setAttribute("cart", map);
+		return "OK";
+	}
+		
 	@RequestMapping("/cart")
 	public String listCart(Model model, HttpServletRequest req) {
 		@SuppressWarnings("unchecked")
@@ -39,6 +40,16 @@ public class CartController {
 //		model.addAttribute("carts", cartDao.listCart());
 //		return "cart/shop_cart";
 		
+	}
+	
+	@GetMapping("/cart/test/add")
+	@ResponseBody
+	public String addCart(MusicVO vo,HttpServletRequest req) {
+		@SuppressWarnings("unchecked")
+		Map<Integer, MusicVO> map = (Map<Integer, MusicVO>) req.getSession().getAttribute("cart");
+		map.put(vo.getId(), vo);
+		req.getSession().setAttribute("cart", map);
+		return "success";
 	}
 	
 	@PostMapping("/deleteCart")
@@ -54,7 +65,15 @@ public class CartController {
 //		} else {
 //		return null;
 		}
-	}
+
+	
+	
+
+//	@GetMapping("/shop/cart")
+//	public String selectCart(CartVO vo, Model model) {
+//		model.addAttribute("cart", cartDao.selectCart(vo));
+//		return "shop/shop_cart";
+//	}	
 	
 //	List<MusicVO> cart = new ArrayList<>();
 //	login할때 
@@ -66,4 +85,4 @@ public class CartController {
 //	(List<MusicVO>) session.getAttribute("cart") 음악 리스트
 //	map.delete(musicVO_Id)
 	
-//}
+}
