+++ src/main/java/com/munhwa/prj/wallet/serviceImpl/UsageServiceImpl.java
@@ -5,6 +5,7 @@ import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
+import com.munhwa.prj.common.vo.Criteria;
 import com.munhwa.prj.wallet.mapper.UsageMapper;
 import com.munhwa.prj.wallet.service.UsageService;
 import com.munhwa.prj.wallet.vo.UsageVO;
@@ -14,8 +15,8 @@ public class UsageServiceImpl implements UsageService {
 	private UsageMapper map;
 	
 	@Override
-	public List<UsageVO> listUsage() {
-		return map.listUsage();
+	public List<UsageVO> findByMemberId(String memberId, Criteria cri) {
+		return map.selectUsageListMemberId(memberId,cri);
 	}
 
 	@Override
@@ -26,6 +27,11 @@ public class UsageServiceImpl implements UsageService {
 	@Override
 	public int insertUsage(UsageVO vo) {
 		return map.insertUsage(vo);
+	}
+
+	@Override
+	public int getCountByUsageId(String id) {
+		return map.getCountByUsageId(id);
 	}
