+++ src/main/webapp/WEB-INF/views/wallet/chargeForm.jsp
@@ -8,7 +8,19 @@
 <script type="text/javascript"
 	src="https://cdn.iamport.kr/js/iamport.payment-1.1.5.js"></script>
 <script>
-   IMP.init('imp66890146');
+var value = document.getElementById('kakaoPay').value
+
+if (value == 'imp66890146') {
+	IMP.init('imp66890146');
+	requestPayKaKao();
+} else ()
+// switch(value){
+// case "imp66890146": {
+// 	IMP.init('imp66890146';)
+// 	requestPayKakao();
+// } 
+}
+   
    function requestPayKaKao() {
       // IMP.request_pay(param, callback) 결제창 호출
       IMP.request_pay({ // param
@@ -34,17 +46,61 @@
     	    		url: "charge",
     	    		type:"post",
     	    		data:{"mileage" : mileage.value,
-    	    			  "memberId" : "test0@gmail.com"
+    	    			  "memberId" : "test1@gmail.com",
+    	    			  "type" : "P01"
     	    		},
+    	    	}).done(function() {
+			    	location.href="walletInfoSelect.do";
     	    	});
     	    } else {
     	        var msg = '결제에 실패하였습니다.';
     	        msg += '에러내용 : ' + rsp.error_msg;
-    	        location.href ="chargeForm";
+    	    	location.href ="chargeForm.do";
     	    }
     	    alert(msg);
-    		location.href="";
     	});
+      
+   }
+   
+   IMP.init('imp04405403');
+   function requestPayCard() {
+      // IMP.request_pay(param, callback) 결제창 호출
+      IMP.request_pay({ // param
+         pg : "smilepay",
+         pay_method : "card",
+         merchant_uid : 'merchant_' + new Date().getTime(),
+         name : "마일리지 충전",
+         amount : mileage.value,
+         buyer_email : "test0@gmail.com",
+<%--         	"<%=(String)session.getAttribute("id")%>", --%>
+         buyer_name : "테스트",
+         buyer_tel : "테스트",
+         buyer_addr : "테스트",
+      }, function(rsp) {
+    	    if ( rsp.success ) {
+    	    	var msg = '결제가 완료되었습니다.';
+//      	        msg += '고유ID : ' + rsp.imp_uid;
+//      	        msg += '상점 거래ID : ' + rsp.merchant_uid;
+//      	        msg += '결제 상품명 : ' + "마일리지 충전";
+     	        msg += '결제 금액 : ' + rsp.paid_amount;
+     	        
+    	    	$.ajax({
+    	    		url: "charge",
+    	    		type:"post",
+    	    		data:{"mileage" : mileage.value,
+    	    			  "memberId" : "test1@gmail.com",
+    	    			  "type" : "P02"
+    	    		},
+    	    	});
+    	    	location.href="walletInfoSelect.do";
+    	    } else {
+    	        var msg = '결제에 실패하였습니다.';
+    	        msg += '에러내용 : ' + rsp.error_msg;
+    	    	location.href ="chargeForm.do";
+    	    }
+    	    alert(msg);
+    	});
+      
    }
 </script>
 <div class="def-block clearfix">
@@ -59,21 +115,24 @@
 		<div align="center"
 			style="color: white; font-size: 20px; margin-bottom: 100px; margin-top: 50px;">
 			충전 수단</div>
-		<button onclick="requestPayKaKao()" style="margin-left: 30px">
+		<button onclick="requestPayKaKao()" style="margin-left: 30px" id="kakaoPay" value="imp66890146">
 			<img src="${path}/resources/images/payment/kakaopay.png" width="150"
 				height="80">
 		</button>
-		<button onclick="requestPayKaKao()" style="margin-left: 30px">
-			<img src="${path}/resources/images/payment/kakaopay.png" width="150"
-				height="80">
-		</button>
-		<button onclick="requestPayKaKao()" style="margin-left: 30px">
-			<img src="${path}/resources/images/payment/kakaopay.png" width="150"
-				height="80">
-		</button>
+		<button onclick="requestPayCard()" style="margin-left: 30px" id="smilePay" value="imp04405403">
+		<img src="${path}/prj/resources/images/payment/kakaopay.png"
+			width="150" height="80">
+	</button>
 	</div>
 	<div align="right">
                 <a href="walletInfo.do" class="tbutton small" style="margin-top:50px"><span>뒤로가기</span></a>
     </div>
+   	<form>
+  		<input type="hidden" value="imp66890146" id="kakao">
+  		<input type="hidden" value="imp04405403" id="smile">
+	</form>
 </div>
 </div>
+<script>
+	
+</script>
