+++ src/main/webapp/WEB-INF/views/wallet/walletInfoSelect.jsp
@@ -5,36 +5,51 @@
 .pageInfo {
 	list-style: none;
 	display: inline-block;
-	margin: 50px 0 0 100px;
+	margin: 10px 0 0 100px;
 }
 
 .pageInfo li {
 	float: left;
-	font-size: 20px;
-	margin-left: 18px;
+	font-size: 1px;
+	margin-left: 10px;
 	padding: 7px;
 	font-weight: 500;
 }
 
 a:link {
-	color: black;
+	color: white;
 	text-decoration: none;
 }
 
 a:visited {
-	color: black;
+	color: white;
 	text-decoration: none;
 }
 
 a:hover {
-	color: black;
+	color: white;
 	text-decoration: underline;
 }
 
-.active{
-      background-color: #cdd5ec;
-  }
- 
+.active {
+	background-color: #cdd5ec;
+}
+
+.search_area {
+	display: inline-block;
+	margin-top: 30px;
+	margin-left: 260px;
+}
+
+.search_area input {
+	height: 30px;
+	width: 250px;
+}
+
+.search_area button {
+	width: 100px;
+	height: 36px;
+}
 </style>
 <div class="def-block clearfix">
 	<div align="right" style="margin-bottom: 50px;">
@@ -45,14 +60,32 @@ a:hover {
 				현재 보유중인 마일리지&nbsp;&nbsp;&nbsp;<input type="text" readonly="readonly"
 					style="height: 15px; margin-top: 5px;">
 			</div>
-<!-- 			<div align="center" style="color: white" class="def-block clearfix"> -->
-<!-- 				요소 선택 <select id="dataPerPage"> -->
-<!-- 					<option value="10">10개씩보기</option> -->
-<!-- 					<option value="15">15개씩보기</option> -->
-<!-- 					<option value="20">20개씩보기</option> -->
-<!-- 				</select> -->
-<!-- 			</div> -->
-
+			<!-- 옵션선택 끝 -->
+			<!-- 			<div class="search_wrap"> -->
+			<!-- 				<div class="search_area"> -->
+			<!-- 					<select name="type"> -->
+			<!-- 						<option value="mileage" -->
+			<%-- 							<c:out value="${pageMaker.cri.type eq 'mileage'?'selected':'' }"/>>가격</option> --%>
+			<!-- 						<option value="name" -->
+			<%-- 							<c:out value="${pageMaker.cri.type eq 'name'?'selected':'' }"/>>결제 수단</option> --%>
+			<!-- 					</select> -->
+			<%-- 					<input type="text" name="keyword" value="${pageMaker.cri.keyword }"> --%>
+			<!-- 					<button>Search</button> -->
+			<!-- 				</div> -->
+			<!-- 			</div> -->
+		</div>
+		<div style="float: right;">
+			<select id="cntPerPage" name="sel" onchange="selChange()">
+				<option value="10"
+					<c:if test="${pageMaker.cri.amount == 10}">selected</c:if>>10줄
+					보기</option>
+				<option value="15"
+					<c:if test="${pageMaker.cri.amount == 15}">selected</c:if>>15줄
+					보기</option>
+				<option value="20"
+					<c:if test="${pageMaker.cri.amount == 20}">selected</c:if>>20줄
+					보기</option>
+			</select>
 		</div>
 		<table class="table">
 			<thead>
@@ -71,11 +104,6 @@ a:hover {
 					</tr>
 				</c:forEach>
 			</tbody>
-			<tfoot>
-				<tr>
-					<td colspan="6">Footer Note: Lorem ipsum dolor sit amet</td>
-				</tr>
-			</tfoot>
 		</table>
 	</div>
 	<div class="pageInfo_wrap">
@@ -83,27 +111,29 @@ a:hover {
 			<ul id="pageInfo" class="pageInfo">
 				<!-- 이전페이지 버튼 -->
 				<c:if test="${pageMaker.prev}">
-					<li class="pageInfo_btn previous"><a
-						href="${pageMaker.startPage-1}">Previous</a></li>
+					<li class="pageInfo_btn previous"><a href="#"
+						onclick="paging(${pageMaker.startPage-1})">Previous</a></li>
 				</c:if>
 				<!-- 각 번호 페이지 버튼 -->
 				<c:forEach var="num" begin="${pageMaker.startPage}"
 					end="${pageMaker.endPage}">
-					<li class="pageInfo_btn ${pageMaker.cri.pageNum == num ? "active":"" }"><button
-						onclick="paging(${num})">${num}</button></li>
+					<li class="pageInfo_btn ${pageMaker.cri.pageNum == num ? "active":"" }"><a
+						href="#" onclick="paging(${num})">${num}</a></li>
 				</c:forEach>
 				<!-- 다음페이지 버튼 -->
 				<c:if test="${pageMaker.next}">
-					<li class="pageInfo_btn next"><a
-						href="${pageMaker.endPage + 1 }">Next</a></li>
+					<li class="pageInfo_btn next"><a href="#"
+						onclick="paging(${pageMaker.endPage + 1})">Next</a></li>
 				</c:if>
 			</ul>
 		</div>
 	</div>
 
-	<form id="moveForm" method="get" action="walletInfoSelect.do" >
+	<form id="moveForm" method="get" action="walletInfoSelect.do">
 		<input type="hidden" name="pageNum" value="${pageMaker.cri.pageNum }">
 		<input type="hidden" name="amount" value="${pageMaker.cri.amount }">
+		<%-- 		<input type="hidden" name="keyword" value="${pageMaker.cri.keyword }"> --%>
+		<%-- 		<input type="hidden" name="type" value="${pageMaker.cri.type }"> --%>
 	</form>
 
 	<div align="right">
@@ -119,5 +149,43 @@ a:hover {
 		moveForm.pageNum.value = num;
 // 		moveForm.attr("action", "walletInfoSelect.do");
 		moveForm.submit();
+		
 	};
+	
+	function selChange() {
+		var sel = document.getElementById('cntPerPage').value;
+		location.href="walletInfoSelect.do?pageNum=1&amount="+sel;
+	}
+
+
+//     $(".search_area button").on("click", function(e){
+//         e.preventDefault();
+//         let val = $("input[name='keyword']").val();
+//         moveForm.find("input[name='keyword']").val(val);
+//         moveForm.find("input[name='pageNum']").val(1);
+//         moveForm.submit();
+//     });
+    
+//     $(".search_area button").on("click", function(e){
+//         e.preventDefault();
+        
+//         let type = $(".search_area select").val();
+//         let keyword = $(".search_area input[name='keyword']").val();
+        
+//         if(!type){
+//             alert("검색 종류를 선택하세요.");
+//             return false;
+//         }
+        
+//         if(!keyword){
+//             alert("키워드를 입력하세요.");
+//             return false;
+//         }        
+        
+//         moveForm.find("input[name='type']").val(type);
+//         moveForm.find("input[name='keyword']").val(keyword);
+//         moveForm.find("input[name='pageNum']").val(1);
+//         moveForm.submit();
+//     });
+	
 </script>
